dev_prueba_viva/requirements.txt
Django
django-filter
django-redis
djangorestframework
djangorestframework-dataclasses
django_mysql
requests